Not finished still trying to understand nested controllers...
I think I have to create another root FXML container, and then children
are pinned to that pane.

Getting this figured out will be confusing, but the good news is that it
will be easy to add components and build upon the interface once that is
done.

So will the login screen be a completely separate thing? Something to
discuss today.