Made changes to login (fixed bug + changed LinkedList to ArrayList).
Created UserRow class. NOTE: It was done this way so that in the future
this can be created server side, and we could bind elements directly to
the table and update them on the fly as they are added.

Created, linked, and populated the user listing table.

TO-DO: Populate edit fields and save changes.